requirements.txt
awscli
pre-commit
detect-secrets
bandit
safety
strands-agents
strands-agents-tools
boto3
botocore

python-dotenv
click
rick

tabulate

pytest
pytest-asyncio
black
flake8

git-remote-codecommit

sympy
matplotlib
requests
